Move Tonic codegen output & converters into new 'pb' module
Reorganise the module hierarchy, placing the autogenerated 'musigrpc'
and 'walletrpc' gRPC bindings under a new 'pb' module, and moving the
marshalling logic ('MyTryInto' trait, impls and 'From' impls converting
between protos and hand-crafted structs) from 'server.rs' to a separate
'crate::pb::convert' module.

(Also inline a few format args in 'server.rs' that were missed by Clippy
for some reason.)
