+++ rpc/src/pb/musigrpc.rs
@@ -1,0 +1,2 @@
+#![allow(clippy::all, clippy::pedantic, clippy::restriction, clippy::nursery)]
+tonic::include_proto!("musigrpc");
